refactor(update-checker): switch from Preact JSX to JS Frontend startup script
